Changed the connection method between the client and server.
First of all, I switched from a UDP connection to a TCP connection because the connection between the server and the client needs to be multithreaded and the connection needs to be maintained. The code for the UDP connection is still there and I will modify it after the lab class on Monday.
Also, when the client connects, it keeps a unique number and sends it to the server so that the server can identify whether the client is disconnected or connected.